cleanup `pmacc/types.hpp`
Cleanup the misc header `pmacc/types.hpp` to support later refactorings of other classes to support compiling in independant compile units.

Move all definitions from types.hpp in independant files.
