+++ include/pmacc/eventSystem/EventType.hpp
@@ -1,4 +1,6 @@
-/* Copyright 2013-2019 Rene Widera
+/* Copyright 2013-2019 Felix Schmitt, Heiko Burau, Rene Widera,
+ *                     Wolfgang Hoenig, Benjamin Worpitz,
+ *                     Alexander Grund
  *
  * This file is part of PMacc.
  *
@@ -21,38 +23,33 @@
 
 #pragma once
 
-#include "pmacc/debug/VerboseLog.hpp"
+#include <cstdint>
 
-#include <stdint.h>
-
-#ifndef PMACC_VERBOSE_LVL
-#define PMACC_VERBOSE_LVL 0
-#endif
 
 namespace pmacc
 {
+namespace eventSystem
+{
 
-    /*create verbose class*/
-    DEFINE_VERBOSE_CLASS(PMaccVerbose)
-    (
-        /* define log lvl for later use
-         * e.g. log<pmaccLogLvl::NOTHING>("TEXT");*/
-        DEFINE_LOGLVL(0,NOTHING);
-        DEFINE_LOGLVL(1,MEMORY);
-        DEFINE_LOGLVL(2,INFO);
-        DEFINE_LOGLVL(4,CRITICAL);
-        DEFINE_LOGLVL(8,MPI);
-        DEFINE_LOGLVL(16,CUDA_RT);
-        DEFINE_LOGLVL(32,COMMUNICATION);
-        DEFINE_LOGLVL(64,EVENT);
-    )
-    /*set default verbose lvl (integer number)*/
-    (NOTHING::lvl|PMACC_VERBOSE_LVL);
+    /**
+     * Internal event/task type used for notifications in the event system.
+     */
+    enum EventType
+    {
+        FINISHED,
+        COPYHOST2DEVICE,
+        COPYDEVICE2HOST,
+        COPYDEVICE2DEVICE,
+        SENDFINISHED,
+        RECVFINISHED,
+        LOGICALAND,
+        SETVALUE,
+        GETVALUE,
+        KERNEL
+    };
 
-    //short name for access verbose types of PMacc
-    using ggLog = PMaccVerbose;
+} // namespace type
 
-}
-
-
-
+    // for backward compatibility pull all definitions into the pmacc namespace
+    using namespace eventSystem;
+} // namespace pmacc
