+++ include/pmacc/type/Integral.hpp
@@ -1,4 +1,6 @@
-/* Copyright 2013-2019 Rene Widera
+/* Copyright 2013-2019 Felix Schmitt, Heiko Burau, Rene Widera,
+ *                     Wolfgang Hoenig, Benjamin Worpitz,
+ *                     Alexander Grund
  *
  * This file is part of PMacc.
  *
@@ -21,38 +23,20 @@
 
 #pragma once
 
-#include "pmacc/debug/VerboseLog.hpp"
+#include <cstdint>
 
-#include <stdint.h>
-
-#ifndef PMACC_VERBOSE_LVL
-#define PMACC_VERBOSE_LVL 0
-#endif
 
 namespace pmacc
 {
+namespace type
+{
 
-    /*create verbose class*/
-    DEFINE_VERBOSE_CLASS(PMaccVerbose)
-    (
-        /* define log lvl for later use
-         * e.g. log<pmaccLogLvl::NOTHING>("TEXT");*/
-        DEFINE_LOGLVL(0,NOTHING);
-        DEFINE_LOGLVL(1,MEMORY);
-        DEFINE_LOGLVL(2,INFO);
-        DEFINE_LOGLVL(4,CRITICAL);
-        DEFINE_LOGLVL(8,MPI);
-        DEFINE_LOGLVL(16,CUDA_RT);
-        DEFINE_LOGLVL(32,COMMUNICATION);
-        DEFINE_LOGLVL(64,EVENT);
-    )
-    /*set default verbose lvl (integer number)*/
-    (NOTHING::lvl|PMACC_VERBOSE_LVL);
+    using id_t = uint64_t;
+    using uint64_cu = unsigned long long int;
+    using int64_cu = long long int;
 
-    //short name for access verbose types of PMacc
-    using ggLog = PMaccVerbose;
+} // namespace type
 
-}
-
-
-
+    // for backward compatibility pull all definitions into the pmacc namespace
+    using namespace type;
+} // namespace pmacc
